+++ crates/deadlock_ref/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ftree_cache"
+name = "deadlock_ref"
 version = "0.2.0"
 authors = ["Maxwell Huang-Hobbs <mhuan13@gmail.com>"]
 edition = "2021"
@@ -13,7 +13,10 @@ serde.workspace = true
 anyhow = { workspace = true }
 swc_common.workspace = true
 parking_lot.workspace = true
-dashmap = "6.0.1"
+dashmap = { version = "6.0.1", optional = true }
 tracing = "0.1.40"
 hashbrown = "0.14.5"
-deadlock_ref = { path = "../deadlock_ref" }
+
+[features]
+default = ["dashmap"]
+dashmap = ["dep:dashmap"]
