+++ crates/deadlock_ref/src/prelude.rs
@@ -1,0 +1,1 @@
+pub use crate::deadlock::{DeadlockDebugRef, MappingRef};
