+++ crates/ftree_cache/src/context_data.rs
@@ -1,15 +1,8 @@
 use anyhow::{anyhow, Context, Error, Result};
 use dashmap::DashMap;
+use deadlock_ref::prelude::*;
 use parking_lot::{MappedRwLockReadGuard, RwLock, RwLockReadGuard, RwLockWriteGuard};
-use std::{
-    backtrace::Backtrace,
-    hash::Hash,
-    ops::Deref,
-    path::{Path, PathBuf},
-    sync::mpsc,
-    thread,
-    time::Duration,
-};
+use std::path::{Path, PathBuf};
 use tracing::debug;
 
 /// Represents some state which is represented by the existence of a path on the filesystem
@@ -170,96 +163,6 @@ pub struct FileContextCache<
     args: TArgs,
 }
 
-struct TimerCanceller {
-    cancel_sender: mpsc::Sender<()>,
-}
-
-impl TimerCanceller {
-    pub fn cancel(&self) {
-        if let Err(e) = self.cancel_sender.send(()) {
-            println!("failed to signal release of DeadlockDebugDroppable! {e}")
-        }
-    }
-}
-
-fn run_with_timer(timeout: Duration, msg: String) -> TimerCanceller {
-    let allocation_trace = Backtrace::capture();
-    let (cancel_sender, cancel_receiver) = mpsc::channel();
-
-    thread::spawn(move || {
-        thread::sleep(timeout);
-        match cancel_receiver.try_recv() {
-            Ok(_) => {} // noop, the object has already been dropped
-            Err(_) => println!(
-                // no message sent before timeout. Warn about deadlock
-                "{msg} timed out after {} ms \
-                Allocation site:\n{allocation_trace}",
-                timeout.as_millis()
-            ),
-        }
-    });
-
-    return TimerCanceller{cancel_sender};
-}
-
-struct DeadlockDebugDroppable {
-    deadlock_warning: TimerCanceller
-}
-
-impl DeadlockDebugDroppable {
-    fn new(timeout: Duration, name: String) -> Self {
-        Self {
-            deadlock_warning: run_with_timer(timeout, format!("release {name}",)),
-        }
-    }
-}
-
-impl Drop for DeadlockDebugDroppable {
-    fn drop(&mut self) {
-        self.deadlock_warning.cancel()
-    }
-}
-
-/* Wrapper reference type that logs when created and logs when released.
- * If it fails to release, maybe there is a deadlock?
- */
-pub struct DeadlockDebugRef<TRef: Deref, const REFERENCE_TIMEOUT_MS: u64 = 5000> {
-    inner_ref: TRef,
-    drop_tracker: DeadlockDebugDroppable,
-}
-
-impl<TRef: Deref, const REFERENCE_TIMEOUT_MS: u64> DeadlockDebugRef<TRef, REFERENCE_TIMEOUT_MS> {
-    fn wrap(to_wrap: TRef, name: String) -> Self {
-        Self {
-            inner_ref: to_wrap,
-            drop_tracker: DeadlockDebugDroppable::new(
-                Duration::from_millis(REFERENCE_TIMEOUT_MS),
-                name,
-            ),
-        }
-    }
-
-    fn create<TInitFn: FnOnce() -> TRef>(name: String, f: TInitFn) -> Self {
-        // Run a potentially deadlocking function, with a timer in parallel to log if it
-        // over-ran the timer.
-        let deadlock_warning = run_with_timer(REFERENCE_TIMEOUT_MS, format!("lock {name}"))
-        let reference = f();
-        deadlock_warning.cancel();
-        Self::wrap(reference, name)
-    }
-}
-
-impl<T, TRef> Deref for DeadlockDebugRef<TRef>
-where
-    TRef: Deref<Target = T>,
-{
-    type Target = T;
-
-    fn deref(&self) -> &Self::Target {
-        return self.inner_ref.deref();
-    }
-}
-
 pub type CtxRef<'a, T> = DeadlockDebugRef<dashmap::mapref::one::Ref<'a, PathBuf, T>>;
 pub type CtxOptRef<'a, T> =
     DeadlockDebugRef<dashmap::mapref::one::MappedRef<'a, PathBuf, Option<T>, T>>;
@@ -270,122 +173,6 @@ impl<T: ContextData<()>, const CONTEXT_FNAME: &'static str> FileContextCache<T, 
             cache: DashMap::new(),
             args: (),
         }
-    }
-}
-
-pub trait MappingRef<T>
-where
-    Self: Sized + Deref<Target = T>,
-{
-    type MappedRef<T2>: Deref<Target = T2>;
-
-    fn map<F, T2>(self, f: F) -> Self::MappedRef<T2>
-    where
-        F: FnOnce(&Self::Target) -> &T2;
-
-    fn try_map<F, T2>(self, f: F) -> Result<Self::MappedRef<T2>, Self>
-    where
-        F: FnOnce(&Self::Target) -> Option<&T2>;
-}
-
-impl<'a, K: Eq + Hash, V> MappingRef<V> for dashmap::mapref::one::Ref<'a, K, V> {
-    type MappedRef<T> = dashmap::mapref::one::MappedRef<'a, K, V, T>;
-
-    fn map<F, T2>(self, f: F) -> Self::MappedRef<T2>
-    where
-        F: FnOnce(&<Self as Deref>::Target) -> &T2,
-    {
-        let me: dashmap::mapref::one::Ref<'a, K, V> = self;
-        me.map(f)
-    }
-
-    fn try_map<F, T2>(self, f: F) -> Result<Self::MappedRef<T2>, Self>
-    where
-        F: FnOnce(&Self::Target) -> Option<&T2>,
-    {
-        let me: dashmap::mapref::one::Ref<'a, K, V> = self;
-        me.try_map(f)
-    }
-}
-
-impl<'a, K: Eq + Hash, V> DeadlockDebugRef<dashmap::mapref::one::Ref<'a, K, V>> {
-    pub fn key(&self) -> &K {
-        self.inner_ref.key()
-    }
-
-    pub fn value(&self) -> &V {
-        self.inner_ref.value()
-    }
-
-    pub fn pair(&self) -> (&K, &V) {
-        self.inner_ref.pair()
-    }
-}
-
-impl<V, TInnerRef: MappingRef<V>> MappingRef<V> for DeadlockDebugRef<TInnerRef> {
-    type MappedRef<T> = DeadlockDebugRef<<TInnerRef as MappingRef<V>>::MappedRef<T>>;
-
-    fn map<F, T2>(self, f: F) -> Self::MappedRef<T2>
-    where
-        F: FnOnce(&<Self as Deref>::Target) -> &T2,
-    {
-        let inner_mapped = self.inner_ref.map(f);
-        DeadlockDebugRef::<<TInnerRef as MappingRef<V>>::MappedRef<T2>> {
-            inner_ref: inner_mapped,
-            drop_tracker: self.drop_tracker,
-        }
-    }
-
-    fn try_map<F, T2>(self, f: F) -> Result<Self::MappedRef<T2>, Self>
-    where
-        F: FnOnce(&Self::Target) -> Option<&T2>,
-    {
-        match self.inner_ref.try_map(f) {
-            Ok(inner_mapped) => Ok(
-                DeadlockDebugRef::<<TInnerRef as MappingRef<V>>::MappedRef<T2>> {
-                    inner_ref: inner_mapped,
-                    drop_tracker: self.drop_tracker,
-                },
-            ),
-            Err(inner_ref) => Err(DeadlockDebugRef::<TInnerRef> {
-                inner_ref,
-                drop_tracker: self.drop_tracker,
-            }),
-        }
-    }
-}
-
-impl<'a, K: Eq + Hash, V, T2> MappingRef<T2> for dashmap::mapref::one::MappedRef<'a, K, V, T2> {
-    type MappedRef<T3> = dashmap::mapref::one::MappedRef<'a, K, V, T3>;
-
-    fn map<F, T3>(self, f: F) -> Self::MappedRef<T3>
-    where
-        F: FnOnce(&<Self as Deref>::Target) -> &T3,
-    {
-        let me: dashmap::mapref::one::MappedRef<'a, K, V, T2> = self;
-        me.map(f)
-    }
-
-    fn try_map<F, T3>(self, f: F) -> Result<Self::MappedRef<T3>, Self>
-    where
-        F: FnOnce(&Self::Target) -> Option<&T3>,
-    {
-        let me: dashmap::mapref::one::MappedRef<'a, K, V, T2> = self;
-        me.try_map(f)
-    }
-}
-
-impl<'a, K: Eq + Hash, V, T2> DeadlockDebugRef<dashmap::mapref::one::MappedRef<'a, K, V, T2>> {
-    pub fn key(&self) -> &K {
-        self.inner_ref.key()
-    }
-
-    pub fn value(&self) -> &T2 {
-        self.inner_ref.value()
-    }
-
-    pub fn pair(&self) -> (&K, &T2) {
-        self.inner_ref.pair()
     }
 }
 
@@ -463,28 +250,31 @@ impl<T: ContextData<TArgs>, TArgs: Copy, const CONTEXT_FNAME: &'static str>
     // filesystem for a tsconfig.json file and caches the result.
     pub fn check_dir<'a>(&'a self, base: &Path) -> Result<CtxRef<'a, Option<T>>, Error> {
         println!("{} check_dir! {}", CONTEXT_FNAME, base.display());
-        let entry = self.cache.entry(base.to_owned());
-        let res = entry.or_try_insert_with(|| {
-            println!(
-                "{} check_dir! - populate - {}",
-                CONTEXT_FNAME,
-                base.display()
-            );
-            let x = self
-                .check_dir_os_fs(base)
-                .with_context(|| format!("Failed while checking {:?} for {}", base, CONTEXT_FNAME));
-            println!(
-                "{} check_dir! - populated! - {}",
-                CONTEXT_FNAME,
-                base.display()
-            );
-            x
-        })?;
-
-        return Ok(DeadlockDebugRef::wrap(
-            res.downgrade(),
+        DeadlockDebugRef::try_create(
             format!("{CONTEXT_FNAME}:{}", base.display()),
-        ));
+            || -> Result<dashmap::mapref::one::Ref<'a, PathBuf, Option<T>>, Error> {
+                let res = self.cache.entry(base.to_owned()).or_try_insert_with(
+                    || -> Result<Option<T>> {
+                        println!(
+                            "{} check_dir! - populate - {}",
+                            CONTEXT_FNAME,
+                            base.display()
+                        );
+                        let x = self.check_dir_os_fs(base).with_context(|| {
+                            format!("Failed while checking {:?} for {}", base, CONTEXT_FNAME)
+                        })?;
+                        println!(
+                            "{} check_dir! - populated! - {}",
+                            CONTEXT_FNAME,
+                            base.display()
+                        );
+                        Ok(x)
+                    },
+                )?;
+                // downgrade mutable ref to non-mutable ref
+                return Ok(res.downgrade());
+            },
+        )
     }
 
     fn check_dir_os_fs(&self, base: &Path) -> Result<Option<T>, Error> {
